Fix CII Best Practices badge info
This fixes the current misleading text about the CII
Best Practices badge. It currently says that "This badge tells us the
repo maintainers are aware of best development practices." - but
merely being "aware" doesn't earn a passing badge.
There's a long list of requirements to earn a passing badge.
Note that this only checks for "passing", not silver or gold.

Fix and clarify the text.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>