📖 Docs: Pinned dependency doc 2 (#1004)
* Modify pinned dependency docs in checks.yaml

The previous changes about pinned dependencies
modified the generated file checks.md, not the source
file checks.yaml. This commit modifies the correct
source file checks.yaml instead. It also tweaks the
text further (while we're at it).

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

* Check in result of make generate-docs

We've modified checks.yaml to improve the pinned
dependency discussion. This checks in the result of
`make generate-docs` so that the docs are visible
on GitHub.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>

Co-authored-by: Azeem Shaikh <azeemshaikh38@gmail.com>