Modify the text to acknowledge GitHub != universe
There are many source repository hosting services (forges),
not just GitHub. This generalizes the requirements, e.g., from:

> Determines if the project's GitHub workflows follow the principle of least privilege.

to:

> Determines if the project's workflows follow the principle of least privilege.

Scorecard doesn't currently *implement* checks in most cases for
systems other than GitHub, so acknowledge that as a limitation, instead
of implying that it's the one true way to implement secure software.

Signed-off-by: David A. Wheeler <dwheeler@dwheeler.com>