fix: Dockerfile and shell script score counted twice
Dependencies insecurely downdloaded and run are identified in different files. These files can be Dockerfiles, shell scripts, etc. When calculating dockerDownloadScore and scriptScore, we were getting  dependencies doing downloadThenRun for all file types. That resulted in penalization for each downloadThenRun twice.
Now, we are treating downloadThenRun as a special case like GHA, and separating the totalization for Dockerfiles and shell scripts.

Signed-off-by: Gabriela Gutierrez <gabigutierrez@google.com>