CODEOWNERS: Support distribution of code reviews via team assignments
Individual maintainer assignments within CODEOWNERS mean that we
cannot take advantage of GitHub code review distribution schemes
for team review assignments.

In this commit, we switch to team assignments within CODEOWNERS.

A common complaint with this approach is that unless you are a part
of the GitHub organization, you will not be able to view a team's
membership/understand who the maintainers of a project are.

To provide visibility into the maintainer list, we've added a
MAINTAINERS.md here as well.

Signed-off-by: Stephen Augustus <foo@auggie.dev>