:seedling: Require minimum version of Go 1.22.0 (#4322)
* Bump go directive to 1.22.0

At least one of our dependencies now requires 1.22. Additionally, with
the Go 1.23 release, Go 1.21 is no longer supported. Due to the loopvar
changes, I'm submitting this as a standalone change.

https://go.dev/doc/go1.22

Signed-off-by: Spencer Schrock <sschrock@google.com>

* remove unneeded nolint directive

This wasn't included in b31449017edbf20ed5b7d3a30bf35c5901062b43,
because we didn't redeclare the loopvar with `tt := tt`, which is now
done by default in Go 1.22.0.

Signed-off-by: Spencer Schrock <sschrock@google.com>

* bump minimum Go version in setup doc

Signed-off-by: Spencer Schrock <sschrock@google.com>

---------

Signed-off-by: Spencer Schrock <sschrock@google.com>