Add a workaround for organisations in GitLab
This fixes Contributors.

Currently organisations are an experimental features in GitLab this change makes
it use the users email domain as a pseudo organisation.  The domains are
anonymised as these are printed as part of the report and shouldn't be exposed.

It's uncertain if the uptake in organisation will be large considering GitLab
users are used to using groups for this purpose.

Signed-off-by: Amar Takhar <amar@rtems.org>