:sparkles: Add GitHub git compatibility mode (#4474)
* add git handler for GitHub repositories

This is primarily aimed at helping in cases where a repository's
.gitattributes file causes files to not be analyzed.

Signed-off-by: Spencer Schrock <sschrock@google.com>

* use variadic options to configure GitHub repoclient

This will let us use the new entrypoint in a backwards compatible way,
similar to the scorecard.Run change made in the v5 release.

Signed-off-by: Spencer Schrock <sschrock@google.com>

* add flag to enable github git mode

Signed-off-by: Spencer Schrock <sschrock@google.com>

* rename flag to be forge agnostic

export-ignore is not a github specific feature, and other forges, like
gitlab, suffer from the same bug.

Signed-off-by: Spencer Schrock <sschrock@google.com>

* move git file handler to internal package

This will allow sharing with GitLab in a followup PR

Signed-off-by: Spencer Schrock <sschrock@google.com>

* add a test

Signed-off-by: Spencer Schrock <sschrock@google.com>

* use new toplevel gitmode argument

also moves a func around for smaller PR diff.

Signed-off-by: Spencer Schrock <sschrock@google.com>

* add path traversal test

Signed-off-by: Spencer Schrock <sschrock@google.com>

* change flag to file-mode

Signed-off-by: Spencer Schrock <sschrock@google.com>

* fix repo typo in options test

the value isn't used to connect to anything though.

Signed-off-by: Spencer Schrock <sschrock@google.com>

---------

Signed-off-by: Spencer Schrock <sschrock@google.com>
Signed-off-by: balteravishay <avishay.balter@gmail.com>