remove scorecard image from Google AR promotion (#4638)
The webhook is used to promote latest (staging) to stable (production)
for the weekly cron, which only includes the remaining 5 images.

We aren't hosting the base scorecard image on GAR, but rather GHCR, and
Cloud Build was disabled for that image already. This causes the webhook
call to return a 500 error as the container doesn't exist for the given
commit.

Signed-off-by: Spencer Schrock <sschrock@google.com>