avoid processing nonexistent codeowner files (#4705)
This worked with the `archive` file mode, because the code returned a
(*os.File)(nil), which could handle the call to Close() on a nil
receiver. However due to how the githubrepo client handled reading files
in `git` mode, nonexistent files returned a truly nil value, with both
the type and value unset. https://go.dev/doc/faq#nil_error

Signed-off-by: Spencer Schrock <sschrock@google.com>